% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot State from Fatality Analysis Reporting System Dataset}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number representing state number.}

\item{year}{A number for which year's datasets should be readed}
}
\value{
This function returns a plot of state from loaded dataset.
If input '\code{state.num}' is not avaliable in the dataset it will
throw an error looking like ""invalid STATE number: ", '\code{state.num}'.
If the '\code{year}' input is not an integer or a class can not
coerced to integer then it would not read the appropriarte file and
would give an error.
}
\description{
This function helps to read dataset in working directory and
plot the state according to state number.
It aimed to plot using datasets from US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System.
}
\examples{
\dontrun{
fars_map_state(50,2014)
}
}
