% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read}
\alias{fars_read}
\title{Read dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character}
}
\value{
The dataset whose name is given as input in tibble data frame format.
If the dataset is not available in the working directory, it gives error.
}
\description{
This function basically helps to read datasets in working directory.
It aimed to read datasets especially from US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System.
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
}
}
