% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read year and month columns from multiple datasets}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of which year's datasets should be readed.}
}
\value{
This function returns a list, year and month columns
of loaded dataframes as elments.  If the datasets are not
available in the working directory, this function gives error like
"warning("invalid year: ", year)".
}
\description{
This function allows you to read month and year columns in more than
one datasets in working directory at once.
It aimed to read datasets especially from US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System.
}
\examples{
\dontrun{
fars_read_years(c(2013,2014))
}
}
