% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read and summarize datasets}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of which year's datasets should be summarized.}
}
\value{
This function returns a dataframe with at leat two columns,
month and year columns which are inputs. Year columns have same names as input
and includes observation numbers per month.
If input is not a integer or a class which can be coerced to integer
this function can not create appropriate file name and can not read
relevant datasets and gives error like "warning("invalid year: ", year)".
}
\description{
This function reads month and year columns in more than
one datasets in working directory at once, merges them and
summarize their count with respect to year and month.
It aimed to summarize datasets especially from US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014))
}
}
